A few GUI updates
The heal spell damages friends though, I have determined that it was
this line in the PlayerActor class:

m.setValue((int)(m.getValue()-(4*Math.sqrt(level))));

in the levelUp() method. I am going to let whoever made that method mess
around with that. There are still a few bugs with the battles that
should get fixed before Wednesday.